% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinfer.R
\name{cinfer}
\alias{cinfer}
\title{Run CINFER to infer mis-segregation rates.}
\usage{
cinfer()
}
\description{
Top level controller script to run CINFER. Handles input data, fetches prior data from CINFERbase, and runs approximate Bayesian computation inference of mis-segregation rates.
}
