% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPrior.R
\name{fetchPrior}
\alias{fetchPrior}
\title{Fetch prior data}
\usage{
fetchPrior(
  steps = "all",
  rate = "all",
  pressure = "all",
  forceDownload = FALSE,
  version = "latest"
)
}
\arguments{
\item{steps}{The range of time steps to include in the prior data. Defaults to "all".}

\item{rate}{The range of mis-segregation rates to include in the prior data. Defaults to "all".}

\item{pressure}{The range of selection pressure magnitudes to include in the prior data. Defaults to "all".}

\item{forceDownload}{Force download the specified CINFERbase version.}

\item{version}{The version of CINFERbase to download and/or use. Defaults to "latest".}
}
\description{
Fetches prior data set from CINFERbase given default or user-specified parameters. Downloads the latest CINFERbase release if not present.
}
